Change uint32 to uint128 in Subscriptions contracts

+++ contracts/mcd/automatic/SubscriptionsProxy.sol
@@ -10,7 +10,7 @@ contract SubscriptionsProxy is ConstantAddresses {
 
     address public constant MONITOR_PROXY_ADDRESS = 0xB77bCacE6Fa6415F40798F9960d395135F4b3cc1;
 
-    function subscribe(uint _cdpId, uint32 _minRatio, uint32 _maxRatio, uint32 _optimalRatioBoost, uint32 _optimalRatioRepay, address _subscriptions) public {
+    function subscribe(uint _cdpId, uint128 _minRatio, uint128 _maxRatio, uint128 _optimalRatioBoost, uint128 _optimalRatioRepay, address _subscriptions) public {
         DSGuard guard = DSGuardFactory(FACTORY_ADDRESS).newGuard();
         DSAuth(address(this)).setAuthority(DSAuthority(address(guard)));
 
@@ -19,7 +19,7 @@ contract SubscriptionsProxy is ConstantAddresses {
         Subscriptions(_subscriptions).subscribe(_cdpId, _minRatio, _maxRatio, _optimalRatioBoost, _optimalRatioRepay);
     }
 
-    function update(uint _cdpId, uint32 _minRatio, uint32 _maxRatio, uint32 _optimalRatioBoost, uint32 _optimalRatioRepay, address _subscriptions) public {
+    function update(uint _cdpId, uint128 _minRatio, uint128 _maxRatio, uint128 _optimalRatioBoost, uint128 _optimalRatioRepay, address _subscriptions) public {
         Subscriptions(_subscriptions).subscribe(_cdpId, _minRatio, _maxRatio, _optimalRatioBoost, _optimalRatioRepay);
     }
